requirements.txt
# torch NOT included (already installed with CUDA)

# Core ML / Transformers
transformers>=4.50.0
accelerate>=0.33.0
datasets>=2.19.0
evaluate>=0.4.2

# PyTorch ecosystem
pytorch-lightning>=1.9.5
torchmetrics>=0.11.4

# Graph Neural Networks
torch-geometric>=2.7.0

# Tokenization / serialization
sentencepiece>=0.1.99
protobuf>=3.20.3

# Core data stack
numpy>=2.1.0
pandas>=2.1.4
scikit-learn>=1.3.2

# Chemistry
rdkit>=2024.9.1
pubchempy>=1.0.4

# Utilities
tensorboard>=2.15.1
rich>=13.7.1
tqdm>=4.66.4
openpyxl>=3.1.2
seaborn>=0.12.2
pydantic>=2.6.4
psutil>=7.2.2
requests>=2.32.5
groq>=1.1.2

# Phase 3 RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0

# Phase 4 RL Detoxification
py2opsin>=1.0.2

# Evaluation (RAGAS framework)
ragas>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0